producers/requirements.txt
kafka-python
lxml
